+++ bin/cli.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+
+const app = require('../out/cli').app;
+app.cli();