+++ src/app/compiler/debug.ts
@@ -7,6 +7,8 @@
 
 import * as fs from 'fs-extra';
 import * as path from 'path';
+import * as luaparse from 'luaparse';
+import * as luamin from 'luamin';
 
 import { env } from '../../env';
 import { ConfigurationType, globals, localize } from '../../globals';
@@ -39,7 +41,8 @@ class DebugCompiler extends BaseCompiler {
         ];
 
         if (!env.config.classic) {
-            files.push(await this.genFile(env.asMapPath(globals.FILE_ENTRY), 'orig' + globals.FILE_ENTRY));
+            const scriptFile = await this.getOriginMapScript();
+            files.push(await this.genFile(scriptFile, 'orig' + globals.FILE_ENTRY));
         }
 
         const code = files.join('\n');
@@ -59,8 +62,20 @@ class DebugCompiler extends BaseCompiler {
     }
 
     async genFile(file: string, name?: string) {
-        const body = this.processCodeMacros(await utils.readFile(file));
-        const comment = this.getCommentEqual(body);
+        let body = this.processCodeMacros(await utils.readFile(file));
+        if (body.indexOf('compiletime') >= 0) {
+            await this.initLuaEngine();
+            const ast = luaparse.parse(body, {
+                locations: true,
+                ranges: true,
+                scope: true,
+                onCreateNode: (node) => {
+                    this.checkCompileTime(file, node);
+                }
+            });
+            body = luamin.minify(ast);
+        }
+        const comment = BaseCompiler.getCommentEqual(body);
         if (!name) {
             name = utils.posixCase(path.relative(env.sourceFolder, file));
         }