+++ webpack.config.js
@@ -8,16 +8,20 @@ const path = require("path");
 const config = {
     target: "node",
 
-    entry: "./src/extension.ts",
+    entry: {
+        extension: "./src/extension.ts",
+        cli: "./src/cli.ts",
+    },
     output: {
         path: path.resolve(__dirname, "out"),
-        filename: "extension.js",
+        filename: "[name].js",
         libraryTarget: "commonjs2",
         devtoolModuleFilenameTemplate: "../[resource-path]"
     },
     devtool: "source-map",
     externals: {
-        vscode: "commonjs vscode"
+        vscode: "commonjs vscode",
+        wasmoon: "./wasmoon",
     },
     resolve: {
         extensions: [".ts", ".js"]